+++ src/styles/PostDropDown.scss
@@ -1,0 +1,7 @@
+
+.post-popup-text {
+    padding-left: 15px;
+}
+.post-popup {
+    border: 1px solid #5E5E5E;
+}
